サバイバーゲームα改 (#44)
* 星評価基準／スコア計算基準の調整

* マジックナンバー撲滅

* エネミーレイヤー

* sprite

* ロックオンカーソルβ版

* インプットサービス

* 永続オブジェクトとインプットシステム

* インプット有効化／無効化

* マスタデータから浮動小数点を排除

* レイヤーマスク定数化

* アイテムドロップ偏り問題修正

* アプリ終了時エラー

* 不要メソッド削除

+++ src/Game.Shared/Runtime/Shared/MasterData/MemoryTables/SurvivorItemMaster.cs
@@ -36,7 +36,7 @@ namespace Game.Library.Shared.MasterData.MemoryTables
         /// <summary>レアリティ（1:コモン, 2:レア, 3:エピック）</summary>
         public int Rarity { get; set; }
 
-        /// <summary>スポーン時のスケール（1.0 = 等倍）</summary>
-        public float Scale { get; set; }
+        /// <summary>スポーン時のスケール（100倍値、ToScale()で変換、100 = 1.0倍）</summary>
+        public int Scale { get; set; }
     }
 }
