+++ src/Game.Shared/Runtime/Shared/MasterData/MemoryTables/SurvivorWeaponLevelMaster.cs
@@ -37,28 +37,28 @@ namespace Game.Library.Shared.MasterData.MemoryTables
         /// <summary>ダメージ量</summary>
         public int Damage { get; set; }
 
-        /// <summary>射程、敵/ターゲット検索範囲</summary>
-        public float Range { get; set; }
+        /// <summary>射程、敵/ターゲット検索範囲（1000倍値、ToUnit()で変換）</summary>
+        public int Range { get; set; }
 
-        /// <summary>生成物の移動速度（units/s）</summary>
+        /// <summary>生成物の移動速度（1000倍値、ToUnit()で変換）</summary>
         public int Speed { get; set; }
 
-        /// <summary>発生時間（ms）</summary>
+        /// <summary>発生時間（ms、ToSeconds()で変換）</summary>
         public int Duration { get; set; }
 
-        /// <summary>再度発動できるまでのクールダウン時間（ms）</summary>
+        /// <summary>再度発動できるまでのクールダウン時間（ms、ToSeconds()で変換）</summary>
         public int Cooldown { get; set; }
 
-        /// <summary>ダメージ発生確率（%）</summary>
+        /// <summary>ダメージ発生確率（万分率、ToRate()で変換）</summary>
         public int ProcRate { get; set; }
 
-        /// <summary>発動の間隔（ms）</summary>
+        /// <summary>発動の間隔（ms、ToSeconds()で変換）</summary>
         public int ProcInterval { get; set; }
 
         /// <summary>1回の発動で生成する数</summary>
         public int EmitCount { get; set; }
 
-        /// <summary>複数生成時の間隔（ms）、0=同時</summary>
+        /// <summary>複数生成時の間隔（ms、ToSeconds()で変換）、0=同時</summary>
         public int EmitDelay { get; set; }
 
         /// <summary>同時存在上限</summary>
@@ -67,24 +67,24 @@ namespace Game.Library.Shared.MasterData.MemoryTables
         /// <summary>生成物のヒット回数（-1=AoE/無限）</summary>
         public int HitCount { get; set; }
 
-        /// <summary>生成物の当たり判定サイズ（%）、100=等倍</summary>
+        /// <summary>生成物の当たり判定サイズ（万分率、ToRate()で変換、10000=等倍）</summary>
         public int HitBoxRate { get; set; }
 
-        /// <summary>クリティカル発生確率（%）</summary>
+        /// <summary>クリティカル発生確率（万分率、ToRate()で変換）</summary>
         public int CritHitRate { get; set; }
 
-        /// <summary>クリティカル倍率（%）、150=1.5倍</summary>
+        /// <summary>クリティカル倍率（万分率、ToRate()で変換、15000=1.5倍）</summary>
         public int CritHitMultiplier { get; set; }
 
         #endregion
 
         #region 武器の軌道・物理特性パラメータ
 
-        /// <summary>ノックバック力</summary>
-        public float Knockback { get; set; }
+        /// <summary>ノックバック力（1000倍値、ToUnit()で変換）</summary>
+        public int Knockback { get; set; }
 
-        /// <summary>引き寄せ力</summary>
-        public float Vacuum { get; set; }
+        /// <summary>引き寄せ力（1000倍値、ToUnit()で変換）</summary>
+        public int Vacuum { get; set; }
 
         /// <summary>回転/周回速度（度/s）</summary>
         public int Spin { get; set; }
@@ -98,7 +98,7 @@ namespace Game.Library.Shared.MasterData.MemoryTables
         /// <summary>チェイン回数、0=チェインなし</summary>
         public int Chain { get; set; }
 
-        /// <summary>追尾性能（%）、0=直進、100=完全追尾</summary>
+        /// <summary>追尾性能（万分率、ToRate()で変換、0=直進、10000=完全追尾）</summary>
         public int Homing { get; set; }
 
         /// <summary>拡散角度（度）、0=集中</summary>
